requirements.txt
discord.py==2.4.0
fastapi==0.115.5
PyGithub==2.5.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
uvicorn==0.32.0
requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
